+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.15.1] - 2023-12-11
+### Fixes
+- Fixed outdated messages, added the `gen_counter_for_pstate` flag
+
+## [0.15.0] - 2023-12-08
+### Changes
+- Various improvements to `awint_dag` for `starlight`
+
 ## [0.14.0] - 2023-12-05
 ### Changes
 - Removed all the old basic epoch structs from `awint_dag` now that they have been implemented in